BuffCy-Minimal/requirements.txt
openai
python-dotenv
pygdbmi